+++ sdk/compute/arm-compute/src/computeManagementClient.ts
@@ -8,40 +8,16 @@
 
 import * as coreClient from "@azure/core-client";
 import * as coreRestPipeline from "@azure/core-rest-pipeline";
+import {
+  PipelineRequest,
+  PipelineResponse,
+  SendRequest
+} from "@azure/core-rest-pipeline";
 import * as coreAuth from "@azure/core-auth";
 import {
-  OperationsImpl,
-  UsageOperationsImpl,
-  VirtualMachineSizesImpl,
-  VirtualMachineScaleSetsImpl,
-  VirtualMachineScaleSetExtensionsImpl,
-  VirtualMachineScaleSetRollingUpgradesImpl,
-  VirtualMachineScaleSetVMExtensionsImpl,
-  VirtualMachineScaleSetVMsImpl,
-  VirtualMachineExtensionsImpl,
-  VirtualMachinesImpl,
-  VirtualMachineImagesImpl,
-  VirtualMachineImagesEdgeZoneImpl,
-  VirtualMachineExtensionImagesImpl,
-  AvailabilitySetsImpl,
-  ProximityPlacementGroupsImpl,
-  DedicatedHostGroupsImpl,
-  DedicatedHostsImpl,
-  SshPublicKeysImpl,
-  ImagesImpl,
-  RestorePointCollectionsImpl,
-  RestorePointsImpl,
-  CapacityReservationGroupsImpl,
-  CapacityReservationsImpl,
-  LogAnalyticsImpl,
-  VirtualMachineRunCommandsImpl,
-  VirtualMachineScaleSetVMRunCommandsImpl,
-  DisksImpl,
-  DiskAccessesImpl,
-  DiskEncryptionSetsImpl,
-  DiskRestorePointOperationsImpl,
-  SnapshotsImpl,
-  ResourceSkusImpl,
+  CommunityGalleriesImpl,
+  CommunityGalleryImagesImpl,
+  CommunityGalleryImageVersionsImpl,
   GalleriesImpl,
   GalleryImagesImpl,
   GalleryImageVersionsImpl,
@@ -50,49 +26,12 @@ import {
   GallerySharingProfileImpl,
   SharedGalleriesImpl,
   SharedGalleryImagesImpl,
-  SharedGalleryImageVersionsImpl,
-  CommunityGalleriesImpl,
-  CommunityGalleryImagesImpl,
-  CommunityGalleryImageVersionsImpl,
-  CloudServiceRoleInstancesImpl,
-  CloudServiceRolesImpl,
-  CloudServicesImpl,
-  CloudServicesUpdateDomainImpl,
-  CloudServiceOperatingSystemsImpl
+  SharedGalleryImageVersionsImpl
 } from "./operations";
 import {
-  Operations,
-  UsageOperations,
-  VirtualMachineSizes,
-  VirtualMachineScaleSets,
-  VirtualMachineScaleSetExtensions,
-  VirtualMachineScaleSetRollingUpgrades,
-  VirtualMachineScaleSetVMExtensions,
-  VirtualMachineScaleSetVMs,
-  VirtualMachineExtensions,
-  VirtualMachines,
-  VirtualMachineImages,
-  VirtualMachineImagesEdgeZone,
-  VirtualMachineExtensionImages,
-  AvailabilitySets,
-  ProximityPlacementGroups,
-  DedicatedHostGroups,
-  DedicatedHosts,
-  SshPublicKeys,
-  Images,
-  RestorePointCollections,
-  RestorePoints,
-  CapacityReservationGroups,
-  CapacityReservations,
-  LogAnalytics,
-  VirtualMachineRunCommands,
-  VirtualMachineScaleSetVMRunCommands,
-  Disks,
-  DiskAccesses,
-  DiskEncryptionSets,
-  DiskRestorePointOperations,
-  Snapshots,
-  ResourceSkus,
+  CommunityGalleries,
+  CommunityGalleryImages,
+  CommunityGalleryImageVersions,
   Galleries,
   GalleryImages,
   GalleryImageVersions,
@@ -101,21 +40,14 @@ import {
   GallerySharingProfile,
   SharedGalleries,
   SharedGalleryImages,
-  SharedGalleryImageVersions,
-  CommunityGalleries,
-  CommunityGalleryImages,
-  CommunityGalleryImageVersions,
-  CloudServiceRoleInstances,
-  CloudServiceRoles,
-  CloudServices,
-  CloudServicesUpdateDomain,
-  CloudServiceOperatingSystems
+  SharedGalleryImageVersions
 } from "./operationsInterfaces";
 import { ComputeManagementClientOptionalParams } from "./models";
 
 export class ComputeManagementClient extends coreClient.ServiceClient {
   $host: string;
   subscriptionId: string;
+  apiVersion: string;
 
   /**
    * Initializes a new instance of the ComputeManagementClient class.
@@ -145,7 +77,7 @@ export class ComputeManagementClient extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-compute/21.0.1`;
+    const packageDetails = `azsdk-js-arm-compute/22.0.0`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
@@ -198,50 +130,12 @@ export class ComputeManagementClient extends coreClient.ServiceClient {
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.operations = new OperationsImpl(this);
-    this.usageOperations = new UsageOperationsImpl(this);
-    this.virtualMachineSizes = new VirtualMachineSizesImpl(this);
-    this.virtualMachineScaleSets = new VirtualMachineScaleSetsImpl(this);
-    this.virtualMachineScaleSetExtensions = new VirtualMachineScaleSetExtensionsImpl(
+    this.apiVersion = options.apiVersion || "2022-08-03";
+    this.communityGalleries = new CommunityGalleriesImpl(this);
+    this.communityGalleryImages = new CommunityGalleryImagesImpl(this);
+    this.communityGalleryImageVersions = new CommunityGalleryImageVersionsImpl(
       this
     );
-    this.virtualMachineScaleSetRollingUpgrades = new VirtualMachineScaleSetRollingUpgradesImpl(
-      this
-    );
-    this.virtualMachineScaleSetVMExtensions = new VirtualMachineScaleSetVMExtensionsImpl(
-      this
-    );
-    this.virtualMachineScaleSetVMs = new VirtualMachineScaleSetVMsImpl(this);
-    this.virtualMachineExtensions = new VirtualMachineExtensionsImpl(this);
-    this.virtualMachines = new VirtualMachinesImpl(this);
-    this.virtualMachineImages = new VirtualMachineImagesImpl(this);
-    this.virtualMachineImagesEdgeZone = new VirtualMachineImagesEdgeZoneImpl(
-      this
-    );
-    this.virtualMachineExtensionImages = new VirtualMachineExtensionImagesImpl(
-      this
-    );
-    this.availabilitySets = new AvailabilitySetsImpl(this);
-    this.proximityPlacementGroups = new ProximityPlacementGroupsImpl(this);
-    this.dedicatedHostGroups = new DedicatedHostGroupsImpl(this);
-    this.dedicatedHosts = new DedicatedHostsImpl(this);
-    this.sshPublicKeys = new SshPublicKeysImpl(this);
-    this.images = new ImagesImpl(this);
-    this.restorePointCollections = new RestorePointCollectionsImpl(this);
-    this.restorePoints = new RestorePointsImpl(this);
-    this.capacityReservationGroups = new CapacityReservationGroupsImpl(this);
-    this.capacityReservations = new CapacityReservationsImpl(this);
-    this.logAnalytics = new LogAnalyticsImpl(this);
-    this.virtualMachineRunCommands = new VirtualMachineRunCommandsImpl(this);
-    this.virtualMachineScaleSetVMRunCommands = new VirtualMachineScaleSetVMRunCommandsImpl(
-      this
-    );
-    this.disks = new DisksImpl(this);
-    this.diskAccesses = new DiskAccessesImpl(this);
-    this.diskEncryptionSets = new DiskEncryptionSetsImpl(this);
-    this.diskRestorePointOperations = new DiskRestorePointOperationsImpl(this);
-    this.snapshots = new SnapshotsImpl(this);
-    this.resourceSkus = new ResourceSkusImpl(this);
     this.galleries = new GalleriesImpl(this);
     this.galleryImages = new GalleryImagesImpl(this);
     this.galleryImageVersions = new GalleryImageVersionsImpl(this);
@@ -251,52 +145,40 @@ export class ComputeManagementClient extends coreClient.ServiceClient {
     this.sharedGalleries = new SharedGalleriesImpl(this);
     this.sharedGalleryImages = new SharedGalleryImagesImpl(this);
     this.sharedGalleryImageVersions = new SharedGalleryImageVersionsImpl(this);
-    this.communityGalleries = new CommunityGalleriesImpl(this);
-    this.communityGalleryImages = new CommunityGalleryImagesImpl(this);
-    this.communityGalleryImageVersions = new CommunityGalleryImageVersionsImpl(
-      this
-    );
-    this.cloudServiceRoleInstances = new CloudServiceRoleInstancesImpl(this);
-    this.cloudServiceRoles = new CloudServiceRolesImpl(this);
-    this.cloudServices = new CloudServicesImpl(this);
-    this.cloudServicesUpdateDomain = new CloudServicesUpdateDomainImpl(this);
-    this.cloudServiceOperatingSystems = new CloudServiceOperatingSystemsImpl(
-      this
-    );
+    this.addCustomApiVersionPolicy(options.apiVersion);
   }
 
-  operations: Operations;
-  usageOperations: UsageOperations;
-  virtualMachineSizes: VirtualMachineSizes;
-  virtualMachineScaleSets: VirtualMachineScaleSets;
-  virtualMachineScaleSetExtensions: VirtualMachineScaleSetExtensions;
-  virtualMachineScaleSetRollingUpgrades: VirtualMachineScaleSetRollingUpgrades;
-  virtualMachineScaleSetVMExtensions: VirtualMachineScaleSetVMExtensions;
-  virtualMachineScaleSetVMs: VirtualMachineScaleSetVMs;
-  virtualMachineExtensions: VirtualMachineExtensions;
-  virtualMachines: VirtualMachines;
-  virtualMachineImages: VirtualMachineImages;
-  virtualMachineImagesEdgeZone: VirtualMachineImagesEdgeZone;
-  virtualMachineExtensionImages: VirtualMachineExtensionImages;
-  availabilitySets: AvailabilitySets;
-  proximityPlacementGroups: ProximityPlacementGroups;
-  dedicatedHostGroups: DedicatedHostGroups;
-  dedicatedHosts: DedicatedHosts;
-  sshPublicKeys: SshPublicKeys;
-  images: Images;
-  restorePointCollections: RestorePointCollections;
-  restorePoints: RestorePoints;
-  capacityReservationGroups: CapacityReservationGroups;
-  capacityReservations: CapacityReservations;
-  logAnalytics: LogAnalytics;
-  virtualMachineRunCommands: VirtualMachineRunCommands;
-  virtualMachineScaleSetVMRunCommands: VirtualMachineScaleSetVMRunCommands;
-  disks: Disks;
-  diskAccesses: DiskAccesses;
-  diskEncryptionSets: DiskEncryptionSets;
-  diskRestorePointOperations: DiskRestorePointOperations;
-  snapshots: Snapshots;
-  resourceSkus: ResourceSkus;
+  /** A function that adds a policy that sets the api-version (or equivalent) to reflect the library version. */
+  private addCustomApiVersionPolicy(apiVersion?: string) {
+    if (!apiVersion) {
+      return;
+    }
+    const apiVersionPolicy = {
+      name: "CustomApiVersionPolicy",
+      async sendRequest(
+        request: PipelineRequest,
+        next: SendRequest
+      ): Promise<PipelineResponse> {
+        const param = request.url.split("?");
+        if (param.length > 1) {
+          const newParams = param[1].split("&").map((item) => {
+            if (item.indexOf("api-version") > -1) {
+              return "api-version=" + apiVersion;
+            } else {
+              return item;
+            }
+          });
+          request.url = param[0] + "?" + newParams.join("&");
+        }
+        return next(request);
+      }
+    };
+    this.pipeline.addPolicy(apiVersionPolicy);
+  }
+
+  communityGalleries: CommunityGalleries;
+  communityGalleryImages: CommunityGalleryImages;
+  communityGalleryImageVersions: CommunityGalleryImageVersions;
   galleries: Galleries;
   galleryImages: GalleryImages;
   galleryImageVersions: GalleryImageVersions;
@@ -306,12 +188,4 @@ export class ComputeManagementClient extends coreClient.ServiceClient {
   sharedGalleries: SharedGalleries;
   sharedGalleryImages: SharedGalleryImages;
   sharedGalleryImageVersions: SharedGalleryImageVersions;
-  communityGalleries: CommunityGalleries;
-  communityGalleryImages: CommunityGalleryImages;
-  communityGalleryImageVersions: CommunityGalleryImageVersions;
-  cloudServiceRoleInstances: CloudServiceRoleInstances;
-  cloudServiceRoles: CloudServiceRoles;
-  cloudServices: CloudServices;
-  cloudServicesUpdateDomain: CloudServicesUpdateDomain;
-  cloudServiceOperatingSystems: CloudServiceOperatingSystems;
 }