+++ sdk/compute/arm-compute/src/operations/communityGalleries.ts
@@ -60,11 +60,11 @@ const getOperationSpec: coreClient.OperationSpec = {
       bodyMapper: Mappers.CloudError
     }
   },
-  queryParameters: [Parameters.apiVersion3],
+  queryParameters: [Parameters.apiVersion],
   urlParameters: [
     Parameters.$host,
     Parameters.subscriptionId,
-    Parameters.location1,
+    Parameters.location,
     Parameters.publicGalleryName
   ],
   headerParameters: [Parameters.accept],