+++ sdk/compute/arm-compute/src/operations/communityGalleryImageVersions.ts
@@ -214,14 +214,14 @@ const getOperationSpec: coreClient.OperationSpec = {
       bodyMapper: Mappers.CloudError
     }
   },
-  queryParameters: [Parameters.apiVersion3],
+  queryParameters: [Parameters.apiVersion],
   urlParameters: [
     Parameters.$host,
     Parameters.subscriptionId,
-    Parameters.location1,
+    Parameters.location,
+    Parameters.publicGalleryName,
     Parameters.galleryImageName,
-    Parameters.galleryImageVersionName,
-    Parameters.publicGalleryName
+    Parameters.galleryImageVersionName
   ],
   headerParameters: [Parameters.accept],
   serializer
@@ -238,13 +238,13 @@ const listOperationSpec: coreClient.OperationSpec = {
       bodyMapper: Mappers.CloudError
     }
   },
-  queryParameters: [Parameters.apiVersion3],
+  queryParameters: [Parameters.apiVersion],
   urlParameters: [
     Parameters.$host,
     Parameters.subscriptionId,
-    Parameters.location1,
-    Parameters.galleryImageName,
-    Parameters.publicGalleryName
+    Parameters.location,
+    Parameters.publicGalleryName,
+    Parameters.galleryImageName
   ],
   headerParameters: [Parameters.accept],
   serializer
@@ -263,10 +263,10 @@ const listNextOperationSpec: coreClient.OperationSpec = {
   urlParameters: [
     Parameters.$host,
     Parameters.subscriptionId,
-    Parameters.nextLink,
-    Parameters.location1,
+    Parameters.location,
+    Parameters.publicGalleryName,
     Parameters.galleryImageName,
-    Parameters.publicGalleryName
+    Parameters.nextLink
   ],
   headerParameters: [Parameters.accept],
   serializer