+++ sdk/compute/arm-compute/src/operations/sharedGalleryImageVersions.ts
@@ -218,11 +218,11 @@ const listOperationSpec: coreClient.OperationSpec = {
       bodyMapper: Mappers.CloudError
     }
   },
-  queryParameters: [Parameters.apiVersion3, Parameters.sharedTo],
+  queryParameters: [Parameters.apiVersion, Parameters.sharedTo],
   urlParameters: [
     Parameters.$host,
     Parameters.subscriptionId,
-    Parameters.location1,
+    Parameters.location,
     Parameters.galleryImageName,
     Parameters.galleryUniqueName
   ],
@@ -241,11 +241,11 @@ const getOperationSpec: coreClient.OperationSpec = {
       bodyMapper: Mappers.CloudError
     }
   },
-  queryParameters: [Parameters.apiVersion3],
+  queryParameters: [Parameters.apiVersion],
   urlParameters: [
     Parameters.$host,
     Parameters.subscriptionId,
-    Parameters.location1,
+    Parameters.location,
     Parameters.galleryImageName,
     Parameters.galleryImageVersionName,
     Parameters.galleryUniqueName
@@ -267,9 +267,9 @@ const listNextOperationSpec: coreClient.OperationSpec = {
   urlParameters: [
     Parameters.$host,
     Parameters.subscriptionId,
-    Parameters.nextLink,
-    Parameters.location1,
+    Parameters.location,
     Parameters.galleryImageName,
+    Parameters.nextLink,
     Parameters.galleryUniqueName
   ],
   headerParameters: [Parameters.accept],