+++ sdk/compute/arm-compute/src/operations/sharedGalleryImages.ts
@@ -183,11 +183,11 @@ const listOperationSpec: coreClient.OperationSpec = {
       bodyMapper: Mappers.CloudError
     }
   },
-  queryParameters: [Parameters.apiVersion3, Parameters.sharedTo],
+  queryParameters: [Parameters.apiVersion, Parameters.sharedTo],
   urlParameters: [
     Parameters.$host,
     Parameters.subscriptionId,
-    Parameters.location1,
+    Parameters.location,
     Parameters.galleryUniqueName
   ],
   headerParameters: [Parameters.accept],
@@ -205,11 +205,11 @@ const getOperationSpec: coreClient.OperationSpec = {
       bodyMapper: Mappers.CloudError
     }
   },
-  queryParameters: [Parameters.apiVersion3],
+  queryParameters: [Parameters.apiVersion],
   urlParameters: [
     Parameters.$host,
     Parameters.subscriptionId,
-    Parameters.location1,
+    Parameters.location,
     Parameters.galleryImageName,
     Parameters.galleryUniqueName
   ],
@@ -230,8 +230,8 @@ const listNextOperationSpec: coreClient.OperationSpec = {
   urlParameters: [
     Parameters.$host,
     Parameters.subscriptionId,
+    Parameters.location,
     Parameters.nextLink,
-    Parameters.location1,
     Parameters.galleryUniqueName
   ],
   headerParameters: [Parameters.accept],